+++ src/Schema/Directives/MorphManyDirective.php
@@ -26,9 +26,8 @@ directive @morphMany(
 
   """
   Allows to resolve the relation as a paginated list.
-  Allowed values: `paginator`, `connection`.
   """
-  type: String
+  type: MorphManyType
 
   """
   Allow clients to query paginated lists without specifying the amount of items.
@@ -49,6 +48,21 @@ directive @morphMany(
   """
   edgeType: String
 ) on FIELD_DEFINITION
+
+"""
+Options for the `type` argument of `@morphMany`.
+"""
+enum MorphManyType {
+    """
+    Offset-based pagination, similar to the Laravel default.
+    """
+    PAGINATOR
+
+    """
+    Cursor-based pagination, compatible with the Relay specification.
+    """
+    CONNECTION
+}
 GRAPHQL;
     }
 }